UserService/requirements.txt
fastapi
pymongo
uvicorn
python-dotenv
requests
aiocache